requirements.txt
python-dotenv==1.0.1 # For reading environment variables stored in .env file
langchain==0.2.2 #gave me an error but i changed the version of python to 3.12
langchain==0.3.0 # updated version March.20.2025
# langchain-community==0.2.3
langchain-community==0.3.0 # updated version March.20.2025
langchain-openai==0.1.8 # For embeddings
# unstructured==0.14.4 this gave me error 'No matching distribution found for unstructured==0.14.4'
unstructured==0.17.0 # Document loading
# onnxruntime==1.17.1 # chromadb dependency: on Mac use `conda install onnxruntime -c conda-forge`
# For Windows users, install Microsoft Visual C++ Build Tools first
# install onnxruntime before installing `chromadb`
chromadb==0.5.0 # Vector storage
openai==1.31.1 # For embeddings
tiktoken==0.7.0  # For embeddings

# install markdown depenendies with: `pip install "unstructured[md]"` after install the requirements file. Leave this line commented out.
# pip install langchain-huggingface
# pip install -U langchain-chroma
